+++ backend/src/bundles/avatar-videos/types/types.ts
@@ -1,0 +1,6 @@
+export {
+    type GetAvatarVideoRequestDto,
+    type GetAvatarVideoResponseDto,
+    type RenderAvatarResponseDto,
+    type RenderAvatarVideoRequestDto,
+} from 'shared';
