+++ backend/src/common/services/azure-ai/apis/avatar-video/libs/types/render-avatar-video-api-argument.type.ts
@@ -1,0 +1,8 @@
+import { type RenderAvatarVideoApiRequestDto } from './types.js';
+
+type RenderAvatarVideoApiArgument = {
+    id: string;
+    payload: RenderAvatarVideoApiRequestDto;
+};
+
+export { type RenderAvatarVideoApiArgument };
