+++ backend/src/common/services/azure-ai/apis/avatar-video/libs/types/render-avatar-video-api-request-dto.type.ts
@@ -1,0 +1,16 @@
+import { type AvatarConfig, type InputKind } from './types.js';
+
+type RenderAvatarVideoApiRequestDto = {
+    inputKind: InputKind;
+    synthesisConfig: {
+        voice: string;
+    };
+    inputs: [
+        {
+            content: string;
+        },
+    ];
+    avatarConfig: AvatarConfig;
+};
+
+export { type RenderAvatarVideoApiRequestDto };
