+++ backend/src/common/services/azure-ai/apis/avatar-video/libs/types/render-avatar-video-arguments.type.ts
@@ -1,4 +1,4 @@
-import { type RenderAvatarVideoRequestDto } from '~/common/services/azure-ai/types/types.js';
+import { type RenderAvatarVideoRequestDto } from './types.js';
 
 type RenderAvatarVideoArgument = {
     id: string;
