+++ backend/src/common/services/azure-ai/apis/avatar-video/libs/types/types.ts
@@ -1,1 +1,12 @@
+export { type RenderAvatarVideoApiArgument } from './render-avatar-video-api-argument.type.js';
+export { type RenderAvatarVideoApiRequestDto } from './render-avatar-video-api-request-dto.type.js';
+export { type RenderAvatarVideoApiResponseDto } from './render-avatar-video-api-response-dto.type.js';
 export { type RenderAvatarVideoArgument } from './render-avatar-video-arguments.type.js';
+export { type GetAvatarVideoResponseDto } from 'shared';
+export {
+    type AvatarConfig,
+    type InputKind,
+    type RenderAvatarVideoRequestDto,
+    type VideoCodec,
+    type VideoFormat,
+} from 'shared';
