--- backend/src/common/services/azure-ai/types/render-avatar-video-request-dto.type.ts
@@ -1,17 +1,0 @@
-import { type AvatarConfig } from './avatar-config.type.js';
-import { type InputKind } from './input-kind.type.js';
-
-type RenderAvatarVideoRequestDto = {
-    inputKind: InputKind;
-    synthesisConfig: {
-        voice: string;
-    };
-    inputs: [
-        {
-            content: string;
-        },
-    ];
-    avatarConfig: AvatarConfig;
-};
-
-export { type RenderAvatarVideoRequestDto };
