+++ backend/src/common/services/azure-ai/types/types.ts
@@ -1,7 +1,1 @@
 export { type AzureGetVoicesResponseDto } from './azure-get-voices-response-dto.type.js';
-export { type GetAvatarVideoRequestDto } from './get-avatar-video-request-dto.type.js';
-export { type GetAvatarVideoResponseDto } from './get-avatar-video-response-dto.type.js';
-export { type InputKind } from './input-kind.type.js';
-export { type RenderAvatarVideoRequestDto } from './render-avatar-video-request-dto.type.js';
-export { type RenderAvatarVideoResponseDto } from './render-avatar-video-response-dto.type.js';
-export { type VideoCodec, type VideoFormat } from 'shared';
