--- shared/src/bundles/avatar-video/avatar-video.ts
@@ -1,2 +1,0 @@
-export { type VideoFormat } from './types/types.js';
-export { type VideoCodec } from './types/types.js';
