--- shared/src/bundles/avatar-video/types/types.ts
@@ -1,2 +1,0 @@
-export { type VideoCodec } from './video-codec.type.js';
-export { type VideoFormat } from './video-format.type.js';
