+++ shared/src/bundles/avatar-videos/types/avatar-config.type.ts
