+++ shared/src/bundles/avatar-videos/types/get-avatar-video-request-dto.type.ts
