+++ shared/src/bundles/avatar-videos/types/get-avatar-video-response-dto.type.ts
