+++ shared/src/bundles/avatar-videos/types/input-kind.type.ts
