+++ shared/src/bundles/avatar-videos/types/render-avatar-request-dto.type.ts
@@ -1,0 +1,8 @@
+type RenderAvatarVideoRequestDto = {
+    avatarName: string;
+    avatarStyle: string;
+    text: string;
+    voice: string;
+};
+
+export { type RenderAvatarVideoRequestDto };
