+++ shared/src/bundles/avatar-videos/types/render-avatar-response-dto.type.ts
@@ -1,0 +1,5 @@
+type RenderAvatarResponseDto = {
+    id: string;
+};
+
+export { type RenderAvatarResponseDto };
