+++ shared/src/bundles/avatar-videos/types/video-codec.type.ts
