+++ shared/src/bundles/avatar-videos/types/video-format.type.ts
